Module Snarky.Snark

include module type of struct include Snarky_backendless.Snark0 end

val set_eval_constraints : bool -> unit

Sets the eval_constraints state. If true, run_unchecked and prove will check that the constraint system is satisfied while evaluating the Checked.t. The default value is false.

Note: This flag will not evaluate any constraints added by with_constraint_system (or the underlying Types.Checked.With_constraint_system). For these, you should modify your code to use the normal run_and_check function.

exception Runtime_error of string list * exn * string

The exception raised when evaluating a checked computation raises an exception.

The arguments are: 1. the backtrace of all active labels at the point in the checked computation when the exception was raised 2. the original exception that was raised 3. the backtrace of the original exception

module Make (Backend : Snarky_backendless.Backend_intf.S) : Snarky_backendless.Snark_intf.S with type field = Backend.Field.t and type Bigint.t = Backend.Bigint.R.t and type R1CS_constraint_system.t = Backend.R1CS_constraint_system.t and type Field.Vector.t = Backend.Field.Vector.t

module Run : sig ... end

type 'field m = (module Snarky_backendless.Snark_intf.Run with type field = 'field)

val make : (module Snarky_backendless.Backend_intf.S with type Field.t = 'field) -> 'field m

module Backend_intf = Snarky_backendless.Backend_intf

module Merkle_tree = Snarky_backendless.Merkle_tree

module Knapsack = Snarky_backendless.Knapsack