Module Snark.Run

val throw_on_id : int -> unit

throw_on_id id set an internal flag that causes Make to throw an error if its internal id would be the same as id.

This can be used to identify where different instances come from, so that the same instance can be used for creating and calling functions.

module Make (Backend : Snarky_backendless.Backend_intf.S) (Prover_state : sig ... end) : Snarky_backendless.Snark_intf.Run with type field = Backend.Field.t and type prover_state = Prover_state.t and type Bigint.t = Backend.Bigint.R.t and type R1CS_constraint_system.t = Backend.R1CS_constraint_system.t and type Var.t = Backend.Var.t and type Field.Constant.Vector.t = Backend.Field.Vector.t and type Verification_key.t = Backend.Verification_key.t and type Proving_key.t = Backend.Proving_key.t