Module Make.Boolean

Representation of booleans within a field.

This representation ties the value of true to Field.one and false to Field.zero, adding a check in Boolean.typ to ensure that these are the only vales.

type var = Field.t Snarky_backendless.Snark_intf.Boolean0.t

type value = bool

val true_ : var

val false_ : var

val if_ : var -> then_:var -> else_:var -> var

val not : var -> var

val (&&) : var -> var -> var

val (&&&) : var -> var -> var

val (||) : var -> var -> var

val (|||) : var -> var -> var

val (lxor) : var -> var -> var

val any : var list -> var

val all : var list -> var

val of_field : Field.t -> var

Convert a value in a field to a boolean, adding checks to the R1CS that it is a valid boolean value.

val var_of_value : value -> var

val typ : (var, value) Typ.t

The relationship between var and value, with a check that the value is valid (ie. Field.zero or Field.one).

val typ_unchecked : (var, value) Typ.t

typ without a validity check for the underlying field value.

val equal : var -> var -> var

module Expr : sig ... end

module Unsafe : sig ... end

module Assert : sig ... end

module Array : sig ... end