Module T.Provider

type nonrec ('a, 'f) t = ( ( 'a Request.t, 'f ) t, ( 'a, 'f ) t ) Types.Provider.t

The different ways to generate a value of type 'a for a circuit witness over field 'f.

This is one of: * a Request, dispatching an 'a Request.t; * Compute, running a computation to generate the value; * Both, attempting to dispatch an 'a Request.t, and falling back to the computation if the request is unhandled or raises an exception.

val run : ( 'a -> 'b Snarky_backendless__Request.t, 'c -> 'd ) Types.Provider.provider -> string list -> 'e -> Request.Handler.t -> 'f