Module 1-Impl.Proof_system

The interface for managing proof systems.

type ('a, 's, 'public_input) t

A proof system instance for a checked computation producing a value of type 'a, with prover state 's and public inputs 'public_input.

val create : ?proving_key:Proving_key.t -> ?verification_key:Verification_key.t -> ?proving_key_path:string -> ?verification_key_path:string -> ?keys_with_hashes:bool -> ?handlers:Handler.t list -> ?reduce:bool -> public_input:(('a, 's) Checked.t, unit, 'computation, 'public_input) Data_spec.t -> 'computation -> ('a, 's, 'public_input) t

Create a new proof system. The arguments are

proving_key -- optional, defines the key to be used for proving. If not present, a key will be read from proving_key_path, or one will be generated automatically.

verification_key -- optional, defines the key to be used for verification of a proof. If not present, a key will be read from verification_key_path, or one will be generated automatically.

proving_key_path -- optional, defines the path to a file where the proving key can be found. If the file does not exist and no proving_key argument is given, the generated key will be written to this file.

verification_key_path -- optional, defines the path to a file where the verification key can be found. If the file does not exist and no verification_key argument is given, the generated key will be written to this file.

keys_with_hashes determines whether keys read from and written to the proving_key_path and verification_key_path should include a MD5 digest of the constraint system. Default value: true.

handlers -- optional, the list of handlers that should be used to handle requests made from the checked computation

reduce -- optional, default false, whether to perform the reduce_to_prover optimisation while creating the proof system

public_input -- the Data_spec.t that describes the form that the public inputs must take

'computation -- a checked computation that takes as arguments values with the types described by public_input to the output type.

val constraint_system : ('a, 's, 'public_input) t -> R1CS_constraint_system.t

The constraint system that this proof system's checked computation describes.

val digest : ('a, 's, 'public_input) t -> Md5_lib.t

The MD5 hash of the constraint system.

val generate_keypair : ('a, 's, 'public_input) t -> Keypair.t

Generate a keypair for the checked computation, writing it to the proving_key_path and verification_key_path, if set.

val run_unchecked : public_input:(unit, 'public_input) H_list.t -> ?handlers:Handler.t list -> ?reduce:bool -> ('a, 's, 'public_input) t -> ('a -> ('b, 's) As_prover.t) -> 's -> 's * 'b

Run the checked computation as the prover, without checking any constraints.

run_unchecked ~public_input proof_system eval prover_state runs the checked computation described by proof_system with public input public_input, then evaluates the result using eval. eval may be used to convert proof system variables back into OCaml values; see As_prover for the available functions.

val run_checked : public_input:(unit, 'public_input) H_list.t -> ?handlers:Handler.t list -> ?reduce:bool -> ('a, 's, 'public_input) t -> ('a -> ('b, 's) As_prover.t) -> 's -> ('s * 'b) Core_kernel.Or_error.t

Run the checked computation as the prover, checking any constraints.

run_checked ~public_input proof_system eval prover_state runs the checked computation described by proof_system with public input public_input, then evaluates the result using eval. eval may be used to convert proof system variables back into OCaml values; see As_prover for the available functions.

val check : public_input:(unit, 'public_input) H_list.t -> ?handlers:Handler.t list -> ?reduce:bool -> ('a, 's, 'public_input) t -> 's -> unit Core_kernel.Or_error.t

Run the checked computation as the prover, returning Ok () if all of the constraints are correct, or an error describing which constraint was not satisfied.

val generate_witness : public_input:(unit, 'public_input) H_list.t -> ?handlers:Handler.t list -> ?reduce:bool -> ('a, 's, 'public_input) t -> 's -> Proof_inputs.t

Generate a witness (auxiliary input) for the given public input.

Returns a record of field vectors {public_inputs; auxiliary_inputs}, corresponding to the given public input and generated auxiliary input.