Module Internal_Basic.Bitstring_checked

Utility functions for dealing with lists of bits in the R1CS.

type t = Boolean.var list

val equal : t -> t -> Boolean.var Checked.t

val equal_expect_true : t -> t -> Boolean.var Checked.t

Equivalent to equal, but avoids computing field elements to represent chunks of the list when not necessary.

NOTE: This will do extra (wasted) work before falling back to the behaviour of equal when the values are not equal.

val lt_value : Boolean.var Bitstring_lib.Bitstring.Msb_first.t -> bool Bitstring_lib.Bitstring.Msb_first.t -> Boolean.var Checked.t

module Assert : sig ... end