Module Typ.Store

A 'value Store.t value describes storing Field.ts in variables for use in the R1CS. It is a monad, which lets us combine these variables to create more complex values.

For example, we can store a 3-tuple of values by writing

let store3 (store_a : 'a Store.t) (store_b : 'b Store.t)
  (store_c : 'b Store.t) : ('a, 'b, 'c) Store.t =
  let open Store in
  let%map a = store_a
  and b = store_b
  and c = store_c
  in
  (a, b, c)

include Monad_let.S with type 'a t = ('a, Field.t) Typ_monads.Store.t

type 'a t = ('a, Field.t) Typ_monads.Store.t

include Core_kernel.Monad.S_without_syntax with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

val all_ignore : unit t list -> unit t

module Let_syntax : sig ... end

val store : field -> Field.Var.t t

Store a single field element for the R1CS, and return the variable that refers to it.