Parameter Make.2-Hash

type var

type value

val typ : (var, value) Impl.Typ.t

val merge : height:int -> var -> var -> var

val if_ : Impl.Boolean.var -> then_:var -> else_:var -> var

val assert_equal : var -> var -> unit

val prover_state : Impl.prover_state

The prover state to run the checked computations above with. This state will *always* be passed to the above unchanged.

NOTE: This is equivalent to the condition on the monadic interface that the computations are not constrained in their prover-state type: the type is abstract from the perspective of the functions, and so they cannot have any effect on the state.