Module Typ.Read

A 'value Read.t describes reading values back out of the R1CS variables, so that the prover can use them in As_prover blocks. It is a monad, which lets us combine these values to create more complex ones.

For example, we can create a record containing the results of some reads by writing

type ('a, 'b) t = {a : 'a; b : 'b}

let read_t (read_a : 'a Read.t) (read_b : 'b Read.t) : ('a, 'b) t =
  let open Read in
  let%map a = read_a
  and b = read_b
  in
  {a; b}

include Monad_let.S with type 'a t = ('a, Field.t) Typ_monads.Read.t

type 'a t = ('a, Field.t) Typ_monads.Read.t

include Core_kernel.Monad.S_without_syntax with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

val all_ignore : unit t list -> unit t

module Let_syntax : sig ... end

val read : Field.Var.t -> field t

Read the contents of a single R1CS variable.