Module Make.As_prover

Code that can be run by the prover only, using 'superpowers' like looking at the contents of R1CS variables and creating new variables from other OCaml values.

type 'a t = ( Field.Var.t -> Field.t ) -> 'a

An ('a) t value generates a value of type 'a.

This type specialises the As_prover.t type for the backend's particular field and variable type.

type 'a as_prover = 'a t

include Snarky_monad_lib.Monad_let.S with type 'a t := 'a t

include Core_kernel.Monad.S_without_syntax with type 'a t := 'a t

val (>>=) : 'a t -> ( 'a -> 'b t ) -> 'b t

val (>>|) : 'a t -> ( 'a -> 'b ) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:( 'a -> 'b t ) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

module Let_syntax : sig ... end

val map2 : 'a t -> 'b t -> f:( 'a -> 'b -> 'c ) -> 'c t

Combine 2 As_prover.t blocks using another function.

val read_var : Field.Var.t -> field t

Read the contents of a R1CS variable representing a single field element.

val read : ( 'var, 'value ) Typ.t -> 'var -> 'value t

read typ x reads the contents of the R1CS variables in x to create an OCaml variable of type 'value, according to the description given by typ.