Module As_prover.Ref

Opaque references for use by the prover in a checked computation.

type 'a t

A mutable reference to an 'a value, which may be used in checked computations.

val create : (unit -> 'a) as_prover -> 'a t

val get : 'a t -> 'a as_prover

val set : 'a t -> 'a -> unit as_prover