Module Internal_Basic.Checked

Checked computations.

These are the values used to generate an R1CS for a computation.

('ret, 'state) t represents a computation 'state -> 'ret that can be compiled into an R1CS.

We form a monad over this type, which allows us to work inside the checked function to do further checked computations. For example (using monad let-syntax):

let multiply3 (x : Field.Var.t) (y : Field.Var.t) (z : Field.Var.t)
  : (Field.Var.t, _) Checked.t =
  open Checked.Let_syntax in
  let%bind x_times_y = Field.Checked.mul x y in
  Field.Checked.mul x_times_y z

type 'prover_state run_state = ('prover_state, Field.t) Run_state.t

include Monad_let.S2 with type ('a, 's) t = ('a, 's, Field.t) Types.Checked.t

type ('a, 's) t = ('a, 's, Field.t) Types.Checked.t

include Monad_let.S_without_syntax2 with type ('a, 'e) t := ('a, 'e) t

include Core_kernel.Monad.Infix2 with type ('a, 'e) t := ('a, 'e) t

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

val (>>|) : ('a, 'e) t -> ('a -> 'b) -> ('b, 'e) t

module Monad_infix : Core_kernel.Monad.Infix2 with type ('a, 'e) t := ('a, 'e) t

val bind : ('a, 'e) t -> f:('a -> ('b, 'e) t) -> ('b, 'e) t

val return : 'a -> ('a, _) t

val map : ('a, 'e) t -> f:('a -> 'b) -> ('b, 'e) t

val join : (('a, 'e) t, 'e) t -> ('a, 'e) t

val ignore_m : (_, 'e) t -> (unit, 'e) t

val all : ('a, 'e) t list -> ('a list, 'e) t

val all_unit : (unit, 'e) t list -> (unit, 'e) t

val all_ignore : (unit, 'e) t list -> (unit, 'e) t

module Let_syntax : sig ... end

module List : Monad_sequence.S with type ('a, 's) monad := ('a, 's) t and type 'a t = 'a list and type boolean := Boolean.var

module Array : Monad_sequence.S with type ('a, 's) monad := ('a, 's) t and type 'a t = 'a array and type boolean := Boolean.var

type Request.t +=

| Choose_preimage : field * int -> bool list Request.t

Choose_preimage is the request issued by Field.Checked.choose_preimage_var before falling back to its default implementation. You can respond to this request to override the default behaviour.

See Request for more information on requests.