Module Snarky_backendless.Snark

include module type of struct include Snark0 end

val set_eval_constraints : bool -> unit

Sets the eval_constraints state. If true, run_unchecked and prove will check that the constraint system is satisfied while evaluating the Checked.t. The default value is false.

Note: This flag will not evaluate any constraints added by with_constraint_system (or the underlying Types.Checked.With_constraint_system). For these, you should modify your code to use the normal run_and_check function.

val set_reduce_to_prover : bool -> unit

Sets the reduce_to_prover state. If true, the Proof_system interface will run optimised versions of the checked computations whenever possible.

Note: This optimisation pre-evaluates and caches some parts of the checked computation, to speed up subsequent calls. *DO NOT USE* if your checked computation uses mutability outside of As_prover blocks.

exception Runtime_error of string * string list * exn * string

The exception raised when evaluating a checked computation raises an exception.

The arguments are: 1. a generic error message, with a presentation of the backtrace 2. the backtrace of all active labels at the point in the checked computation when the exception was raised 3. the original exception that was raised 4. the backtrace of the original exception

module Make (Backend : Backend_intf.S) : Snark_intf.S with type field = Backend.Field.t and type Bigint.t = Backend.Bigint.R.t and type R1CS_constraint_system.t = Backend.R1CS_constraint_system.t and type Var.t = Backend.Var.t and type Field.Vector.t = Backend.Field.Vector.t and type Verification_key.t = Backend.Verification_key.t and type Proving_key.t = Backend.Proving_key.t

module Run : sig ... end

type 'field m = (module Snark_intf.Run with type field = 'field)

type ('prover_state, 'field) m' = (module Snark_intf.Run with type field = 'field and type prover_state = 'prover_state)

val make : (module Backend_intf.S with type Field.t = 'field) -> (unit, 'field) m'

val make' : (module Backend_intf.S with type Field.t = 'field) -> ('prover_state, 'field) m'

val ignore_state : ('prover_state, 'field) m' -> 'field m

module Backend_intf = Backend_intf

module Merkle_tree = Merkle_tree

module Knapsack = Knapsack