Module Make.As_prover

Code that can be run by the prover only, using 'superpowers' like looking at the contents of R1CS variables and creating new variables from other OCaml values.

type ('a, 'prover_state) t = ('a, field, 'prover_state) Snark_intf.As_prover0.t

An ('a, 'prover_state) t value uses the current 'prover_state to generate a value of type 'a, and update the 'prover_state as necessary, within a checked computation.

This type specialises the As_prover.t type for the backend's particular field and variable type.

type ('a, 'prover_state) as_prover = ('a, 'prover_state) t

module Ref : sig ... end

Mutable references for use by the prover in a checked computation.

include Monad_let.S2 with type ('a, 's) t := ('a, 's) t

include Monad_let.S_without_syntax2 with type ('a, 'e) t := ('a, 'e) t

include Core_kernel.Monad.Infix2 with type ('a, 'e) t := ('a, 'e) t

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

val (>>|) : ('a, 'e) t -> ('a -> 'b) -> ('b, 'e) t

module Monad_infix : Core_kernel.Monad.Infix2 with type ('a, 'e) t := ('a, 'e) t

val bind : ('a, 'e) t -> f:('a -> ('b, 'e) t) -> ('b, 'e) t

val return : 'a -> ('a, _) t

val map : ('a, 'e) t -> f:('a -> 'b) -> ('b, 'e) t

val join : (('a, 'e) t, 'e) t -> ('a, 'e) t

val ignore_m : (_, 'e) t -> (unit, 'e) t

val all : ('a, 'e) t list -> ('a list, 'e) t

val all_unit : (unit, 'e) t list -> (unit, 'e) t

val all_ignore : (unit, 'e) t list -> (unit, 'e) t

module Let_syntax : sig ... end

val map2 : ('a, 's) t -> ('b, 's) t -> f:('a -> 'b -> 'c) -> ('c, 's) t

Combine 2 As_prover.t blocks using another function.

val read_var : Field.Var.t -> (field, 'prover_state) t

Read the contents of a R1CS variable representing a single field element.

val get_state : ('prover_state, 'prover_state) t

Read the 'prover_state carried by the As_prover.t monad.

val set_state : 'prover_state -> (unit, 'prover_state) t

Update the 'prover_state carried by the As_prover.t monad.

val modify_state : ('prover_state -> 'prover_state) -> (unit, 'prover_state) t

Change the 'prover_state carried by the As_prover.t monad.

val read : ('var, 'value) Typ.t -> 'var -> ('value, 'prover_state) t

read typ x reads the contents of the R1CS variables in x to create an OCaml variable of type 'value, according to the description given by typ.

val with_lens : ('whole, 'lens) Lens.t -> ('a, 'lens) t -> ('a, 'whole) t

with_lens lens as_prover uses the Lens.t provided to lift the prover state of as_prover to 'whole from a sub-type 'lens.